Stagging commit
WIP task manager
